rename: 폴더명 파스칼 -> 케밥 네이밍 수정
